Delete old XML classes and resources
