+++ app/src/main/kotlin/at/bitfire/davdroid/ui/AccountsActivity.kt
@@ -87,7 +87,6 @@ import at.bitfire.davdroid.db.AppDatabase
 import at.bitfire.davdroid.servicedetection.RefreshCollectionsWorker
 import at.bitfire.davdroid.syncadapter.SyncUtils
 import at.bitfire.davdroid.syncadapter.SyncWorker
-import at.bitfire.davdroid.ui.account.AccountActivity
 import at.bitfire.davdroid.ui.account.AccountActivity2
 import at.bitfire.davdroid.ui.account.AppWarningsModel
 import at.bitfire.davdroid.ui.intro.IntroActivity
