+++ app/src/main/kotlin/at/bitfire/davdroid/ui/account/AccountActivity.kt
@@ -1,63 +1,93 @@
-/***************************************************************************************************
- * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
- **************************************************************************************************/
-
 package at.bitfire.davdroid.ui.account
 
 import android.accounts.Account
-import android.accounts.AccountManager
-import android.accounts.OnAccountsUpdateListener
-import android.app.Application
 import android.content.Intent
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
-import android.view.Menu
-import android.view.MenuInflater
-import android.view.MenuItem
+import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.widget.TooltipCompat
-import androidx.core.view.MenuProvider
-import androidx.fragment.app.FragmentActivity
-import androidx.lifecycle.AndroidViewModel
-import androidx.lifecycle.MutableLiveData
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.pager.HorizontalPager
+import androidx.compose.foundation.pager.rememberPagerState
+import androidx.compose.material.AlertDialog
+import androidx.compose.material.Checkbox
+import androidx.compose.material.DropdownMenu
+import androidx.compose.material.DropdownMenuItem
+import androidx.compose.material.ExperimentalMaterialApi
+import androidx.compose.material.FloatingActionButton
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
+import androidx.compose.material.LinearProgressIndicator
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.ProgressIndicatorDefaults
+import androidx.compose.material.Scaffold
+import androidx.compose.material.Tab
+import androidx.compose.material.TabRow
+import androidx.compose.material.Text
+import androidx.compose.material.TextButton
+import androidx.compose.material.TopAppBar
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.CreateNewFolder
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.DriveFileRenameOutline
+import androidx.compose.material.icons.filled.MoreVert
+import androidx.compose.material.icons.filled.Settings
+import androidx.compose.material.icons.filled.Sync
+import androidx.compose.material.icons.outlined.RuleFolder
+import androidx.compose.material.pullrefresh.PullRefreshIndicator
+import androidx.compose.material.pullrefresh.pullRefresh
+import androidx.compose.material.pullrefresh.rememberPullRefreshState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.graphics.graphicsLayer
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.viewModelScope
-import androidx.viewpager2.adapter.FragmentStateAdapter
+import androidx.paging.compose.LazyPagingItems
+import androidx.paging.compose.collectAsLazyPagingItems
 import at.bitfire.davdroid.R
-import at.bitfire.davdroid.databinding.ActivityAccountBinding
-import at.bitfire.davdroid.db.AppDatabase
 import at.bitfire.davdroid.db.Collection
-import at.bitfire.davdroid.db.Service
-import at.bitfire.davdroid.log.Logger
+import at.bitfire.davdroid.servicedetection.RefreshCollectionsWorker
 import at.bitfire.davdroid.settings.AccountSettings
 import at.bitfire.davdroid.syncadapter.SyncWorker
-import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import com.google.android.material.snackbar.Snackbar
-import com.google.android.material.tabs.TabLayoutMediator
-import dagger.assisted.Assisted
-import dagger.assisted.AssistedFactory
-import dagger.assisted.AssistedInject
+import com.google.accompanist.themeadapter.material.MdcTheme
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.NonCancellable
 import kotlinx.coroutines.launch
-import java.util.logging.Level
 import javax.inject.Inject
 
 @AndroidEntryPoint
-class AccountActivity: AppCompatActivity() {
+class AccountActivity2 : AppCompatActivity() {
 
     companion object {
         const val EXTRA_ACCOUNT = "account"
     }
 
-    @Inject lateinit var modelFactory: Model.Factory
-    val model by viewModels<Model> {
+    @Inject
+    lateinit var modelFactory: AccountModel.Factory
+    val model by viewModels<AccountModel> {
         val account = intent.getParcelableExtra(EXTRA_ACCOUNT) as? Account
-                ?: throw IllegalArgumentException("AccountActivity requires EXTRA_ACCOUNT")
+            ?: throw IllegalArgumentException("AccountActivity requires EXTRA_ACCOUNT")
         object: ViewModelProvider.Factory {
             @Suppress("UNCHECKED_CAST")
             override fun <T: ViewModel> create(modelClass: Class<T>) =
@@ -65,286 +95,463 @@ class AccountActivity: AppCompatActivity() {
         }
     }
 
-    private val warningsModel by viewModels<AppWarningsModel>()
-
-    private lateinit var binding: ActivityAccountBinding
-
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        title = model.account.name
-
-        binding = ActivityAccountBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        setSupportActionBar(binding.toolbar)
-        supportActionBar?.setDisplayHomeAsUpEnabled(true)
-
-        model.accountExists.observe(this) { accountExists ->
-            if (!accountExists)
+        model.invalid.observe(this) { invalid ->
+            if (invalid)
+                // account does not exist anymore
                 finish()
         }
 
-        model.services.observe(this) { services ->
-            val calDavServiceId = services.firstOrNull { it.type == Service.TYPE_CALDAV }?.id
-            val cardDavServiceId = services.firstOrNull { it.type == Service.TYPE_CARDDAV }?.id
-
-            val viewPager = binding.viewPager
-            val adapter = FragmentsAdapter(this, cardDavServiceId, calDavServiceId)
-            viewPager.adapter = adapter
-
-            // connect ViewPager with TabLayout (top bar with tabs)
-            TabLayoutMediator(binding.tabLayout, viewPager) { tab, position ->
-                tab.text = adapter.getHeading(position)
-            }.attach()
-        }
-
-        // "Sync now" fab
-        TooltipCompat.setTooltipText(binding.sync, binding.sync.contentDescription)
-        warningsModel.networkAvailable.observe(this) { networkAvailable ->
-            binding.sync.setOnClickListener {
-                val msgId =
-                    if (warningsModel.networkAvailable.value == true)
-                        R.string.sync_started
-                    else
-                        R.string.no_internet_sync_scheduled
-                Snackbar.make(
-                    binding.sync,
-                    msgId,
-                    Snackbar.LENGTH_SHORT
-                ).show()
-                SyncWorker.enqueueAllAuthorities(this, model.account)
-            }
-        }
-
-        addMenuProvider(object : MenuProvider {
-            override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
-                menuInflater.inflate(R.menu.activity_account, menu)
-            }
-
-            override fun onMenuItemSelected(menuItem: MenuItem) =
-                when (menuItem.itemId) {
-                    R.id.settings -> {
-                        openAccountSettings()
-                        true
-                    }
-                    R.id.rename_account -> {
-                        renameAccount()
-                        true
-                    }
-                    R.id.delete_account -> {
-                        deleteAccountDialog()
-                        true
-                    }
-                    else -> false
+        setContent {
+            MdcTheme {
+                val cardDavSvc by model.cardDavSvc.observeAsState()
+                val cardDavRefreshing by model.cardDavRefreshingActive.observeAsState(false)
+                val cardDavSyncActive by model.cardDavSyncActive.observeAsState(false)
+                val cardDavSyncPending by model.cardDavSyncPending.observeAsState(false)
+                val cardDavProgress = when {
+                    cardDavRefreshing || cardDavSyncActive -> ServiceProgressValue.ACTIVE
+                    cardDavSyncPending -> ServiceProgressValue.PENDING
+                    else -> ServiceProgressValue.IDLE
                 }
-        })
-    }
+                val addressBooks by model.addressBooksPager.observeAsState()
 
-
-    // menu actions
-
-    fun openAccountSettings() {
-        val intent = Intent(this, SettingsActivity::class.java)
-        intent.putExtra(SettingsActivity.EXTRA_ACCOUNT, model.account)
-        startActivity(intent, null)
-    }
-
-    fun renameAccount() {
-        RenameAccountFragment.newInstance(model.account).show(supportFragmentManager, null)
-    }
-
-    fun deleteAccountDialog() {
-        MaterialAlertDialogBuilder(this)
-                .setIcon(R.drawable.ic_error)
-                .setTitle(R.string.account_delete_confirmation_title)
-                .setMessage(R.string.account_delete_confirmation_text)
-                .setNegativeButton(android.R.string.no, null)
-                .setPositiveButton(android.R.string.yes) { _, _ ->
-                    deleteAccount()
+                val calDavSvc by model.calDavSvc.observeAsState()
+                val calDavRefreshing by model.calDavRefreshingActive.observeAsState(false)
+                val calDavSyncActive by model.calDavSyncActive.observeAsState(false)
+                val calDavSyncPending by model.calDavSyncPending.observeAsState(false)
+                val calDavProgress = when {
+                    calDavRefreshing || calDavSyncActive -> ServiceProgressValue.ACTIVE
+                    calDavSyncPending -> ServiceProgressValue.PENDING
+                    else -> ServiceProgressValue.IDLE
                 }
-                .show()
+                val calendars by model.calendarsPager.observeAsState()
+                val subscriptions by model.webcalPager.observeAsState()
+
+                AccountOverview(
+                    account = model.account,
+                    showOnlyPersonal = model.showOnlyPersonal.observeAsState(AccountSettings.ShowOnlyPersonal(false, true)).value,
+                    onSetShowOnlyPersonal = {
+                        model.setShowOnlyPersonal(it)
+                    },
+                    hasCardDav = cardDavSvc != null,
+                    cardDavRefreshing = cardDavProgress,
+                    addressBooks = addressBooks?.flow?.collectAsLazyPagingItems(),
+                    hasCalDav = calDavSvc != null,
+                    calDavRefreshing = calDavProgress,
+                    calendars = calendars?.flow?.collectAsLazyPagingItems(),
+                    subscriptions = subscriptions?.flow?.collectAsLazyPagingItems(),
+                    onUpdateCollectionSync = { id, sync ->
+                        model.setCollectionSync(id, sync)
+                    },
+                    onChangeForceReadOnly = { id, forceReadOnly ->
+                        model.setCollectionForceReadOnly(id, forceReadOnly)
+                    },
+                    onRefreshCollections = {
+                        cardDavSvc?.let { svc ->
+                            RefreshCollectionsWorker.enqueue(this@AccountActivity2, svc.id)
+                        }
+                        calDavSvc?.let { svc ->
+                            RefreshCollectionsWorker.enqueue(this@AccountActivity2, svc.id)
+                        }
+                    },
+                    onSync = {
+                        SyncWorker.enqueueAllAuthorities(this, model.account)
+                    },
+                    onAccountSettings = {
+                        val intent = Intent(this, SettingsActivity::class.java)
+                        intent.putExtra(SettingsActivity.EXTRA_ACCOUNT, model.account)
+                        startActivity(intent, null)
+                    },
+                    onDeleteAccount = {
+                        model.deleteAccount()
+                    },
+                    onNavUp = ::onNavigateUp
+                )
+            }
+        }
     }
 
-    private fun deleteAccount() {
-        val accountManager = AccountManager.get(this)
+}
 
-        accountManager.removeAccount(model.account, this, { future ->
-            try {
-                if (future.result.getBoolean(AccountManager.KEY_BOOLEAN_RESULT))
-                    Handler(Looper.getMainLooper()).post {
-                        finish()
+
+enum class ServiceProgressValue {
+    IDLE,
+    PENDING,
+    ACTIVE
+}
+
+@OptIn(ExperimentalFoundationApi::class, ExperimentalMaterialApi::class)
+@Composable
+fun AccountOverview(
+    account: Account,
+    showOnlyPersonal: AccountSettings.ShowOnlyPersonal,
+    onSetShowOnlyPersonal: (showOnlyPersonal: Boolean) -> Unit,
+    hasCardDav: Boolean,
+    cardDavRefreshing: ServiceProgressValue,
+    addressBooks: LazyPagingItems<Collection>?,
+    hasCalDav: Boolean,
+    calDavRefreshing: ServiceProgressValue,
+    calendars: LazyPagingItems<Collection>?,
+    subscriptions: LazyPagingItems<Collection>?,
+    onUpdateCollectionSync: (collectionId: Long, sync: Boolean) -> Unit = { _, _ -> },
+    onChangeForceReadOnly: (collectionId: Long, forceReadOnly: Boolean) -> Unit = { _, _ -> },
+    onRefreshCollections: () -> Unit = {},
+    onSync: () -> Unit = {},
+    onAccountSettings: () -> Unit = {},
+    onDeleteAccount: () -> Unit = {},
+    onNavUp: () -> Unit = {}
+) {
+    val context = LocalContext.current
+
+    val refreshing by remember { mutableStateOf(false) }
+    val pullRefreshState = rememberPullRefreshState(
+        refreshing,
+        onRefresh = onRefreshCollections
+    )
+    // TODO pull with collection colors
+
+    var showDeleteAccountDialog by remember { mutableStateOf(false) }
+
+    // tabs calculation
+    var currentIdx = -1
+    @Suppress("KotlinConstantConditions")
+    val idxCardDav: Int? = if (hasCardDav) ++currentIdx else null
+    val idxCalDav: Int? = if (hasCalDav) ++currentIdx else null
+    val idxWebcal: Int? = if ((subscriptions?.itemCount ?: 0) > 0) ++currentIdx else null
+    val nrPages =
+        (if (idxCardDav != null) 1 else 0) +
+                (if (idxCalDav != null) 1 else 0) +
+                (if (idxWebcal != null) 1 else 0)
+    val pagerState = rememberPagerState(pageCount = { nrPages })
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                navigationIcon = {
+                    IconButton(onClick = onNavUp) {
+                        Icon(
+                            Icons.AutoMirrored.Default.ArrowBack,
+                            contentDescription = stringResource(R.string.navigate_up)
+                        )
                     }
-            } catch(e: Exception) {
-                Logger.log.log(Level.SEVERE, "Couldn't remove account", e)
-            }
-        }, null)
-    }
+                },
+                title = {
+                    Text(
+                        account.name,
+                        maxLines = 1,
+                        overflow = TextOverflow.Ellipsis
+                    )
+                },
+                actions = {
+                    var overflowOpen by remember { mutableStateOf(false) }
+                    IconButton(onClick = onAccountSettings) {
+                        Icon(Icons.Default.Settings, stringResource(R.string.account_settings))
+                    }
+                    IconButton(onClick = { overflowOpen = !overflowOpen }) {
+                        Icon(Icons.Default.MoreVert, null)
+                    }
+                    DropdownMenu(
+                        expanded = overflowOpen,
+                        onDismissRequest = { overflowOpen = false }
+                    ) {
+                        // TAB-SPECIFIC ACTIONS
 
+                        // create address book
+                        if (pagerState.currentPage == idxCardDav) {
+                            // create address book
+                            DropdownMenuItem(onClick = {
+                                val intent = Intent(context, CreateAddressBookActivity::class.java)
+                                intent.putExtra(CreateAddressBookActivity.EXTRA_ACCOUNT, account)
+                                context.startActivity(intent)
 
-    // public functions
+                                overflowOpen = false
+                            }) {
+                                Icon(
+                                    Icons.Default.CreateNewFolder, stringResource(R.string.create_addressbook),
+                                    modifier = Modifier.padding(end = 8.dp)
+                                )
+                                Text(stringResource(R.string.create_addressbook))
+                            }
+                        } else if (pagerState.currentPage == idxCalDav) {
+                            // create calendar
+                            DropdownMenuItem(onClick = {
+                                val intent = Intent(context, CreateCalendarActivity::class.java)
+                                intent.putExtra(CreateCalendarActivity.EXTRA_ACCOUNT, account)
+                                context.startActivity(intent)
 
-    /**
-     * Updates the click listener of the refresh collections list FAB, according to the given
-     * fragment. Should be called when the related fragment is resumed.
-     */
-    fun updateRefreshCollectionsListAction(fragment: CollectionsFragment) {
-        val label = when (fragment) {
-            is AddressBooksFragment ->
-                getString(R.string.account_refresh_address_book_list)
+                                overflowOpen = false
+                            }) {
+                                Icon(
+                                    Icons.Default.CreateNewFolder, stringResource(R.string.create_calendar),
+                                    modifier = Modifier.padding(end = 8.dp)
+                                )
+                                Text(stringResource(R.string.create_calendar))
+                            }
+                        }
 
-            is CalendarsFragment,
-            is WebcalFragment ->
-                getString(R.string.account_refresh_calendar_list)
+                        // GENERAL ACTIONS
 
-            else -> null
-        }
-        if (label != null) {
-            binding.refresh.contentDescription = label
-            TooltipCompat.setTooltipText(binding.refresh, label)
-        }
+                        // show only personal
+                        DropdownMenuItem(
+                            onClick = {
+                                onSetShowOnlyPersonal(!showOnlyPersonal.onlyPersonal)
+                                overflowOpen = false
+                            },
+                            enabled = !showOnlyPersonal.locked
+                        ) {
+                            Text(stringResource(R.string.account_only_personal))
+                            Checkbox(
+                                checked = showOnlyPersonal.onlyPersonal,
+                                enabled = !showOnlyPersonal.locked,
+                                onCheckedChange = {
+                                    onSetShowOnlyPersonal(it)
+                                    overflowOpen = false
+                                }
+                            )
+                        }
 
-        binding.refresh.setOnClickListener {
-            fragment.onRefresh()
-        }
-    }
+                        // rename account
+                        DropdownMenuItem(onClick = {
+                            /* rename account */
+                        }) {
+                            Icon(
+                                Icons.Default.DriveFileRenameOutline, stringResource(R.string.account_rename),
+                                modifier = Modifier.padding(end = 8.dp)
+                            )
+                            Text(stringResource(R.string.account_rename))
+                        }
 
-
-
-    // adapter
-
-    class FragmentsAdapter(
-        val activity: FragmentActivity,
-        private val cardDavSvcId: Long?,
-        private val calDavSvcId: Long?
-    ): FragmentStateAdapter(activity) {
-
-        private val idxCardDav: Int?
-        private val idxCalDav: Int?
-        private val idxWebcal: Int?
-
-        init {
-            var currentIndex = 0
-
-            idxCardDav = if (cardDavSvcId != null)
-                currentIndex++
-            else
-                null
-
-            if (calDavSvcId != null) {
-                idxCalDav = currentIndex++
-                idxWebcal = currentIndex
-            } else {
-                idxCalDav = null
-                idxWebcal = null
-            }
-        }
-
-        override fun getItemCount() =
-            (if (idxCardDav != null) 1 else 0) +
-            (if (idxCalDav != null) 1 else 0) +
-            (if (idxWebcal != null) 1 else 0)
-
-        override fun createFragment(position: Int) =
-            when (position) {
-                idxCardDav ->
-                    AddressBooksFragment().apply {
-                        arguments = Bundle(2).apply {
-                            putLong(CollectionsFragment.EXTRA_SERVICE_ID, cardDavSvcId!!)
-                            putString(CollectionsFragment.EXTRA_COLLECTION_TYPE, Collection.TYPE_ADDRESSBOOK)
+                        // delete account
+                        DropdownMenuItem(onClick = {
+                            showDeleteAccountDialog = true
+                            overflowOpen = false
+                        }) {
+                            Icon(
+                                Icons.Default.Delete, stringResource(R.string.account_delete),
+                                modifier = Modifier.padding(end = 8.dp)
+                            )
+                            Text(stringResource(R.string.account_delete))
                         }
                     }
-                idxCalDav ->
-                    CalendarsFragment().apply {
-                        arguments = Bundle(2).apply {
-                            putLong(CollectionsFragment.EXTRA_SERVICE_ID, calDavSvcId!!)
-                            putString(CollectionsFragment.EXTRA_COLLECTION_TYPE, Collection.TYPE_CALENDAR)
-                        }
-                    }
-                idxWebcal ->
-                    WebcalFragment().apply {
-                        arguments = Bundle(2).apply {
-                            putLong(CollectionsFragment.EXTRA_SERVICE_ID, calDavSvcId!!)
-                            putString(CollectionsFragment.EXTRA_COLLECTION_TYPE, Collection.TYPE_WEBCAL)
-                        }
-                    }
-                else -> throw IllegalArgumentException()
+
+                    if (showDeleteAccountDialog)
+                        DeleteAccountDialog(
+                            onConfirm = onDeleteAccount,
+                            onDismiss = { showDeleteAccountDialog = false }
+                        )
+                }
+            )
+        },
+        floatingActionButton = {
+            Column {
+                FloatingActionButton(
+                    onClick = onRefreshCollections,
+                    backgroundColor = MaterialTheme.colors.background,
+                    modifier = Modifier.padding(bottom = 16.dp)
+                ) {
+                    Icon(Icons.Outlined.RuleFolder, stringResource(R.string.account_refresh_collections))
+                }
+
+                FloatingActionButton(onClick = onSync) {
+                    Icon(Icons.Default.Sync, stringResource(R.string.account_synchronize_now))
+                }
             }
+        },
+        modifier = Modifier.pullRefresh(pullRefreshState)
+    ) { padding ->
+        Column {
+            if (nrPages > 0) {
+                val scope = rememberCoroutineScope()
 
-        fun getHeading(position: Int) =
-            when (position) {
-                idxCardDav -> activity.getString(R.string.account_carddav)
-                idxCalDav -> activity.getString(R.string.account_caldav)
-                idxWebcal -> activity.getString(R.string.account_webcal)
-                else -> throw IllegalArgumentException()
-            }
+                TabRow(
+                    selectedTabIndex = pagerState.currentPage,
+                    modifier = Modifier.padding(padding)
+                ) {
+                    if (idxCardDav != null)
+                        Tab(
+                            selected = pagerState.currentPage == idxCardDav,
+                            onClick = {
+                                scope.launch {
+                                    pagerState.scrollToPage(idxCardDav)
+                                }
+                            }
+                        ) {
+                            Text(
+                                stringResource(R.string.account_carddav).uppercase(),
+                                modifier = Modifier.padding(8.dp)
+                            )
+                        }
 
-    }
+                        AnimatedVisibility(idxCalDav != null) {
+                            Tab(
+                                selected = pagerState.currentPage == idxCalDav,
+                                onClick = {
+                                    if (idxCalDav != null)
+                                        scope.launch {
+                                            pagerState.scrollToPage(idxCalDav)
+                                        }
+                                }
+                            ) {
+                                Text(
+                                    stringResource(R.string.account_caldav).uppercase(),
+                                    modifier = Modifier.padding(8.dp)
+                                )
+                            }
+                        }
 
+                        AnimatedVisibility(idxWebcal != null) {
+                            Tab(
+                                selected = pagerState.currentPage == idxWebcal,
+                                onClick = {
+                                    if (idxWebcal != null)
+                                        scope.launch {
+                                            pagerState.scrollToPage(idxWebcal)
+                                        }
+                                }
+                            ) {
+                                Text(
+                                    stringResource(R.string.account_webcal).uppercase(),
+                                    modifier = Modifier.padding(8.dp)
+                                )
+                            }
+                        }
+                }
 
-    // model
+                HorizontalPager(
+                    pagerState,
+                    verticalAlignment = Alignment.Top,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .weight(1f)
+                ) { index ->
+                    Box {
+                        when (index) {
+                            idxCardDav ->
+                                ServiceTab(
+                                    refreshing = cardDavRefreshing,
+                                    collections = addressBooks,
+                                    onUpdateCollectionSync = onUpdateCollectionSync,
+                                    onChangeForceReadOnly = onChangeForceReadOnly
+                                )
 
-    class Model @AssistedInject constructor(
-        application: Application,
-        val db: AppDatabase,
-        @Assisted val account: Account
-    ): AndroidViewModel(application), OnAccountsUpdateListener {
+                            idxCalDav ->
+                                ServiceTab(
+                                    refreshing = calDavRefreshing,
+                                    collections = calendars,
+                                    onUpdateCollectionSync = onUpdateCollectionSync,
+                                    onChangeForceReadOnly = onChangeForceReadOnly
+                                )
 
-        @AssistedFactory
-        interface Factory {
-            fun create(account: Account): Model
-        }
+                            idxWebcal ->
+                                ServiceTab(
+                                    refreshing = calDavRefreshing,
+                                    collections = subscriptions,
+                                    onUpdateCollectionSync = onUpdateCollectionSync,
+                                    onChangeForceReadOnly = onChangeForceReadOnly
+                                )
+                        }
 
-        val accountManager: AccountManager = AccountManager.get(application)
-        val accountSettings by lazy { AccountSettings(application, account) }
-
-        val accountExists = MutableLiveData<Boolean>()
-        val services = db.serviceDao().getServiceTypeAndIdsByAccount(account.name)
-
-        val showOnlyPersonal = MutableLiveData<Boolean>()
-        val showOnlyPersonalWritable = MutableLiveData<Boolean>()
-
-
-        init {
-            accountManager.addOnAccountsUpdatedListener(this, null, true)
-            viewModelScope.launch(Dispatchers.IO) {
-                accountSettings.getShowOnlyPersonalPair().let { (value, locked) ->
-                    showOnlyPersonal.postValue(value)
-                    showOnlyPersonalWritable.postValue(locked)
+                        PullRefreshIndicator(
+                            refreshing = refreshing,
+                            state = pullRefreshState,
+                            modifier = Modifier.align(Alignment.TopCenter)
+                        )
+                    }
                 }
             }
         }
-
-        override fun onCleared() {
-            accountManager.removeOnAccountsUpdatedListener(this)
-        }
-
-        override fun onAccountsUpdated(accounts: Array<out Account>) {
-            accountExists.postValue(accounts.contains(account))
-        }
-
-        fun toggleReadOnly(item: Collection) {
-            viewModelScope.launch(Dispatchers.IO + NonCancellable) {
-                val newItem = item.copy(forceReadOnly = !item.forceReadOnly)
-                db.collectionDao().update(newItem)
-            }
-        }
-
-        fun toggleShowOnlyPersonal() {
-            showOnlyPersonal.value?.let { oldValue ->
-                val newValue = !oldValue
-                accountSettings.setShowOnlyPersonal(newValue)
-                showOnlyPersonal.postValue(newValue)
-            }
-        }
-
-        fun toggleSync(item: Collection) {
-            viewModelScope.launch(Dispatchers.IO + NonCancellable) {
-                val newItem = item.copy(sync = !item.sync)
-                db.collectionDao().update(newItem)
-            }
-        }
-
     }
+}
 
+@Preview
+@Composable
+fun AccountOverview_CardDAV_CalDAV() {
+    AccountOverview(
+        account = Account("test@example.com", "test"),
+        showOnlyPersonal = AccountSettings.ShowOnlyPersonal(false, true),
+        onSetShowOnlyPersonal = {},
+        hasCardDav = true,
+        cardDavRefreshing = ServiceProgressValue.ACTIVE,
+        addressBooks = null,
+        hasCalDav = true,
+        calDavRefreshing = ServiceProgressValue.PENDING,
+        calendars = null,
+        subscriptions = null
+    )
+}
+
+@Composable
+@Preview
+fun DeleteAccountDialog(
+    onConfirm: () -> Unit = {},
+    onDismiss: () -> Unit = {}
+) {
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text(stringResource(R.string.account_delete_confirmation_title)) },
+        text = { Text(stringResource(R.string.account_delete_confirmation_text)) },
+        confirmButton = {
+            TextButton(onClick = onConfirm) {
+                Text(stringResource(android.R.string.yes).uppercase())
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDismiss) {
+                Text(stringResource(android.R.string.no).uppercase())
+            }
+        }
+    )
+}
+
+@Composable
+fun ServiceTab(
+    refreshing: ServiceProgressValue,
+    collections: LazyPagingItems<Collection>?,
+    onUpdateCollectionSync: (collectionId: Long, sync: Boolean) -> Unit,
+    onChangeForceReadOnly: (collectionId: Long, forceReadOnly: Boolean) -> Unit
+) {
+    Column {
+        // progress indicator
+        val progressAlpha by animateFloatAsState(
+            when (refreshing) {
+                ServiceProgressValue.ACTIVE -> 1f
+                ServiceProgressValue.PENDING -> .5f
+                else -> 0f
+            },
+            label = "cardDavProgress"
+        )
+        when (refreshing) {
+            ServiceProgressValue.ACTIVE ->
+                // indeterminate
+                LinearProgressIndicator(
+                    color = MaterialTheme.colors.secondary,
+                    modifier = Modifier
+                        .graphicsLayer(alpha = progressAlpha)
+                        .fillMaxWidth()
+                )
+            ServiceProgressValue.PENDING ->
+                // determinate 100%, but semi-transparent (see progressAlpha)
+                LinearProgressIndicator(
+                    color = MaterialTheme.colors.secondary,
+                    progress = 1f,
+                    modifier = Modifier
+                        .alpha(progressAlpha)
+                        .fillMaxWidth()
+                )
+            else ->
+                Spacer(Modifier.height(ProgressIndicatorDefaults.StrokeWidth))
+        }
+
+        //  collection list
+        if (collections != null)
+            CollectionsList(
+                collections,
+                onChangeSync = onUpdateCollectionSync,
+                onChangeForceReadOnly = onChangeForceReadOnly,
+                modifier = Modifier.weight(1f)
+            )
+    }
 }
