--- app/src/main/res/layout/account_caldav_item.xml
@@ -1,100 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              xmlns:tools="http://schemas.android.com/tools"
-              xmlns:app="http://schemas.android.com/apk/res-auto"
-              android:orientation="horizontal"
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content"
-              android:paddingStart="0dp"
-              android:paddingEnd="8dp">
-
-    <View
-        android:id="@+id/color"
-        android:layout_width="8dp"
-        android:layout_height="match_parent"
-        android:layout_marginEnd="4dp"
-        tools:background="@color/primaryColor"/>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center_vertical"
-        android:paddingTop="4dp"
-        android:paddingBottom="4dp"
-        android:orientation="horizontal">
-
-        <com.google.android.material.switchmaterial.SwitchMaterial
-            android:id="@+id/sync"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:focusable="false"
-            android:focusableInTouchMode="false"
-            android:clickable="false"
-            android:contentDescription="@string/account_synchronize_this_collection"
-            android:layout_marginEnd="4dp"/>
-
-        <LinearLayout
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:orientation="vertical">
-
-            <TextView
-                android:id="@+id/title"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textAlignment="viewStart"
-                android:textAppearance="?android:attr/textAppearanceMedium"
-                tools:text="My Calendar" />
-
-            <TextView
-                android:id="@+id/description"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:textAlignment="viewStart"
-                android:textAppearance="?android:attr/textAppearanceSmall"
-                tools:text="Calendar Description" />
-
-        </LinearLayout>
-
-        <ImageView
-            android:id="@+id/read_only"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:contentDescription="@string/account_read_only"
-            app:srcCompat="@drawable/ic_remove_circle" />
-
-        <ImageView
-            android:id="@+id/events"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:contentDescription="@string/account_calendar"
-            app:srcCompat="@drawable/ic_today"/>
-
-        <ImageView
-            android:id="@+id/tasks"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:contentDescription="@string/account_task_list"
-            app:srcCompat="@drawable/ic_playlist_add_check"/>
-
-        <ImageView
-            android:id="@+id/journals"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:contentDescription="@string/account_journal"
-            app:srcCompat="@drawable/ic_journals"/>
-
-        <ImageView
-            android:id="@+id/action_overflow"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:paddingLeft="0dp"
-            android:paddingRight="0dp"
-            android:paddingStart="0dp"
-            android:paddingEnd="0dp"
-            style="@style/Widget.AppCompat.ActionButton.Overflow"/>
-
-    </LinearLayout>
-
-</LinearLayout>
