--- app/src/main/res/layout/account_carddav_item.xml
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-              xmlns:app="http://schemas.android.com/apk/res-auto"
-              xmlns:tools="http://schemas.android.com/tools"
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content"
-              android:paddingTop="4dp"
-              android:paddingBottom="4dp"
-              android:paddingStart="8dp"
-              android:paddingEnd="8dp"
-              android:gravity="center_vertical">
-
-    <com.google.android.material.switchmaterial.SwitchMaterial
-        android:id="@+id/sync"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="4dp"
-        android:focusable="false"
-        android:focusableInTouchMode="false"
-        android:clickable="false" />
-
-    <LinearLayout android:layout_width="0dp"
-                  android:layout_weight="1"
-                  android:layout_height="wrap_content"
-                  android:orientation="vertical">
-
-        <TextView
-            android:id="@+id/title"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textAlignment="viewStart"
-            android:textAppearance="?android:attr/textAppearanceMedium"
-            tools:text="My Address Book" />
-
-        <TextView
-            android:id="@+id/description"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textAlignment="viewStart"
-            android:textAppearance="?android:attr/textAppearanceSmall"
-            tools:text="Address Book Description" />
-
-    </LinearLayout>
-
-    <ImageView
-        android:id="@+id/read_only"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:contentDescription="@string/account_read_only"
-        app:srcCompat="@drawable/ic_remove_circle"/>
-
-    <ImageView
-        android:id="@+id/action_overflow"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:paddingEnd="0dp"
-        style="@style/Widget.AppCompat.ActionButton.Overflow"/>
-
-</LinearLayout>
