--- app/src/main/res/menu/account_collection_operations.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item android:id="@+id/force_read_only"
-          android:checkable="true"
-          android:title="@string/collection_force_read_only"/>
-
-    <item android:id="@+id/properties"
-        android:title="@string/collection_properties"/>
-
-    <item android:id="@+id/delete_collection"
-        android:title="@string/delete_collection"/>
-
-</menu>
