--- app/src/main/res/menu/activity_account.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-      xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <group android:orderInCategory="100">
-
-        <item android:id="@+id/settings"
-              android:icon="@drawable/ic_settings"
-              android:title="@string/account_settings"
-              app:showAsAction="ifRoom"/>
-
-        <item android:id="@+id/rename_account"
-              android:title="@string/account_rename"
-              app:showAsAction="never"/>
-
-        <item android:id="@+id/delete_account"
-              android:title="@string/account_delete"
-              app:showAsAction="never"/>
-
-    </group>
-
-</menu>
