--- app/src/main/res/menu/caldav_actions.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item android:id="@+id/showOnlyPersonal"
-        android:title="@string/account_only_personal"
-        android:checkable="true" />
-
-    <item android:id="@+id/refresh"
-          android:title="@string/account_refresh_calendar_list" />
-
-    <item android:id="@+id/create_calendar"
-          android:title="@string/account_create_new_calendar"/>
-
-</menu>
