--- app/src/main/res/menu/carddav_actions.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item android:id="@+id/showOnlyPersonal"
-        android:title="@string/account_only_personal"
-        android:checkable="true" />
-
-    <item android:id="@+id/refresh"
-        android:title="@string/account_refresh_address_book_list"
-        app:showAsAction="never"/>
-
-    <item android:id="@+id/create_address_book"
-        android:title="@string/account_create_new_address_book"
-        app:showAsAction="never"/>
-
-</menu>
